Fix Form Submit incorrect query https://bugs.webkit.org/show_bug.cgi?id=243595
Reviewed by Ryosuke Niwa.

When we submit a form and call submit() again in JavaScript code
that was called, for example, from onsubmit, we lose this submit
event. This happens because some control flags were activated
in the normal submit event and the form was already processing a
submit event generated by the user clicking the submit button.

To solve this problem, the change is:

- Checking if the form submit event is coming from JavaScript code and allowing the event to complete and reset the flags related with form submit.

* Source/WebCore/html/HTMLFormElement.cpp:
(WebCore::HTMLFormElement::submit):

* LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-double-submit-preventdefault-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-double-submit-requestsubmit-expected.txt:

Canonical link: https://commits.webkit.org/285498@main