Enable library validation for webpushd https://bugs.webkit.org/show_bug.cgi?id=281869 rdar://137709472
Reviewed by Elliott Williams.

The webpushd that is bundled in Safari updates for downlevel macOS uses the wrong version of
WebKit. Instead of using the WebKit bundled in the update, it uses the system WebKit. This breaks
Web Push since Safari tries to send messages to webpushd using a newer IPC version than the one that
webpushd's system WebKit expects.

The reason for this is that library validation isn't enabled on webpushd, which is a prerequisite
for webpushd picking the updated WebKit rather than the system WebKit.

* Source/WebKit/Configurations/webpushd.xcconfig:

Canonical link: https://commits.webkit.org/285571@main