Sometimes, WKTR installation does not succeed in the Simulator https://bugs.webkit.org/show_bug.cgi?id=281875 rdar://138351968
Reviewed by Elliott Williams.

This is caused by a daemon involved in the installation being suspended. This patch works around this
issue by creating a small utility that sends a XPC message to the daemon in order to resume execution.
This issue is flaky, so it is hard to confirm the fix is sufficient, but local testing looks good.

* Tools/Scripts/webkitpy/xcode/simulated_device.py:
(SimulatedDevice.install_app):
* Tools/WebKitTestRunner/Configurations/WebKitEligibilityUtil.xcconfig: Added.
* Tools/WebKitTestRunner/WebKitEligibilityUtil/main.cpp: Added.
(main):
* Tools/WebKitTestRunner/WebKitTestRunner.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/285637@main