[JSC] Remove outdated FIXME in JSDOMConvertDate.h https://bugs.webkit.org/show_bug.cgi?id=282067
Reviewed by Geoffrey Garen.

In JSDOMConvertDate there's a fixme that the convert function should be
taking a JSDOMGlobalObject and passing it to jsDate.

This is confusing because it's already taking a JSC::JSGlobalObject&,
and it's unclear why a JSDOMGlobalObject would be better.

The reason for the FIXME appears to be that this function used to accept
a JSC::ExecState& state - but in bug 202392 / commit
https://commits.webkit.org/216662@main we switched away from ExecState.
So it looks like we just forgot to remove the FIXME when we made the
relevant change.

* Source/WebCore/bindings/js/JSDOMConvertDate.h:

Canonical link: https://commits.webkit.org/285746@main