webkitpy.test.markers.xfail(reason=...) throws https://bugs.webkit.org/show_bug.cgi?id=281870
Reviewed by Jonathan Bedard.

We need to actually quote the dict keys.

While we're at it, based on the recommendation of
https://docs.pytest.org/en/7.2.x/example/simple.html#detect-if-running-from-within-a-pytest-run
to detect whether or not we're running under pytest, rather than just
checking whether or not the pytest module can be imported.

* Tools/Scripts/webkitpy/test/markers.py:
(xfail):
(xfail.decorator):
(slow):

Canonical link: https://commits.webkit.org/285777@main