[css-flexbox] incorrect layout with max-height and flex-direction column and justify-content center https://bugs.webkit.org/show_bug.cgi?id=282036
Reviewed by NOBODY (OOPS!).

After calculation of the flex item size and adjusting it to min/max size
of the flex item we should also check how the calculated size
corresponds to specified max size of the whole container. In case the
adjusted flex item size is bigger than the max size of the whole
container we should use the flex item size calculated before adjusting
to flex item min/max size.

There is new layout test which tests the problem.

* LayoutTests/css3/flexbox/nested-flexbox-max-height-expected.txt: Added.
* LayoutTests/css3/flexbox/nested-flexbox-max-height.html: Added.
* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::resolveFlexibleLengths):