Clean up permissions for declarativeNetRequestFeedback and getMatchedRules https://bugs.webkit.org/show_bug.cgi?id=282260 rdar://137267655
Reviewed by Timothy Hatcher.

Without this change, there was no way for an extension with activeTab +
declarativeNetRequestFeedback to grant permission to the URLs that would be included in
getMatchedRules.

This PR also passes tabs to some of the permissions checks in declarativeNetRequestGetMatchedRules
in case a first party load is blocked.

* Source/WebKit/UIProcess/Extensions/Cocoa/API/WebExtensionContextAPIDeclarativeNetRequestCocoa.mm:
(WebKit::WebExtensionContext::declarativeNetRequestGetMatchedRules):
* Source/WebKit/UIProcess/Extensions/Cocoa/WebExtensionContextCocoa.mm:
(WebKit::WebExtensionContext::permissionState):

Canonical link: https://commits.webkit.org/285861@main